Fix quit handling
Each factory now keeps a reference to the protocol it created. That way
the bot can access individual connections and send a quit message to
each.

The quit itself is a bit of a hack. I couldn't figure out how to make
sure the 'quit' messages are actually sent to the server before
reactor.stop() is executed. This custom sigint handler + calling
reactor.stop() with a delay seems to solve it for now